Container-Root/src/experimental/faiss_deduplicate/requirements.txt
argparse
pandas
pyarrow
faiss-gpu
matplotlib
